SDK - Moved the @python_component decorator test to dsl tests
